backend/requirements.txt
fastapi
uvicorn[standard]
yfinance
polygon-api-client
feedparser
numpy
scipy
pandas
python-dotenv
httpx
aiohttp
